requirements.txt
langchain==0.3.6
langchain-huggingface==0.1.0
langchain-community==0.3.4
streamlit==1.39.0 
chromadb==0.5.3
pypdf==4.3.1
sentence-transformers==3.1.1
python-dotenv==1.0.1 
huggingface_hub==0.23.5 
transformers==4.46.3 
torch==2.7.0
pdfplumber==0.11.4 
faiss-cpu==1.8.0